+++ src/main/java/raisetech/StudentManagement/StudentManagementApplication.java
@@ -5,27 +5,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentsCourses;
+import raisetech.StudentManagement.repository.StudentRepository;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
-
-  @Autowired
-  private StudentRepository repository;
-
 
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
-  @GetMapping("/studentList")
-  public List<Student> getStudentList(){
-    return repository.search();
-  }
 
-  @GetMapping("/students-courses")
-  public List<StudentsCourses> getAllCourses(){
-    return repository.allCourses();
-  }
 }
