+++ src/main/java/raisetech/StudentManagement/repository/StudentRepository.java
@@ -1,17 +1,19 @@
-package raisetech.StudentManagement;
+package raisetech.StudentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentsCourses;
 
 @Mapper
 public interface StudentRepository {
 
   @Select("SELECT * FROM students")
-  List<Student> search();
+  List<Student> getSearch();
 
   @Select("SELECT * FROM students_courses")
-  List<StudentsCourses> allCourses();
+  List<StudentsCourses> getAllCourses();
